processor/requirements.txt
kafka-python==2.0.2
elasticsearch==8.4.3
